Added ability to create a graph from a grid.
DFS currently will crash for relatively large grids - 75x75 works but we
need a way to create a grid with barriors.

This would mean to we need to look at the Grid or the DFS to ensure we
can create some marked vertex initially.

DFS with no barries has ALOT of paths that can be travelled - not very
efficient.

The Graph Constructor can connect N,S,E,W as well as using the diagonals
is you wish by setting a boolean.